content header added to reduce redundancy
