+++ actionlist/app/components/action_item_header.js
@@ -59,4 +59,3 @@ ActionItemHeader.propTypes = {
 };
 
 export default ActionItemHeader;
-
